[Docs] Rest Docs 적용 시작 - build.gradle 변경 - UserController 생성
